时序预测库

+++ models/FBNet.py
@@ -36,7 +36,7 @@ class Model(nn.Module):
         """
         """
         # 获取所有时间特征的索引，避免多次long()操作
-        # time_marks = time_marks.long()  # 将所有时间标记转化为长整型
+        time_marks = time_marks.long()  # 将所有时间标记转化为长整型
         #  B,L
         year_mark, quarter_mark, month_mark, week_mark, day_mark, hour_mark = time_marks[:, :, 0], time_marks[:, :, 1], time_marks[:, :, 2], time_marks[:, :, 3], time_marks[:, :, 4], time_marks[:, :, 5]
 
@@ -55,6 +55,7 @@ class Model(nn.Module):
         # x_mark_enc: [Batch, Input length, 5 or 6 ] 对应年月日周时分
         # 输入数据 以及他的时间编码 96*7  预测 192*7
         # 拼接batch_x_mark和batch_y_mark，形成新的张量
+
         combined_time_marks = torch.cat((batch_x_mark, batch_y_mark), dim=1)  # (B, L_x + L_y, 6)
         # 获取拼接后的时间特征的趋势
         combined_trends = self.get_trends(combined_time_marks)
@@ -68,7 +69,7 @@ class Model(nn.Module):
         trends_x = trends_x + self.bias  # (B, L_x, channels)
         trends_y = trends_y + self.bias  # (B, L_y, channels)
         batch_x = batch_x  - trends_x  # (B, L_x, channels)
-        y_pred    = self.model(batch_x.permute(0, 2, 1)) .permute(0, 2, 1) # (B, L_y, pred_len)
+        y_pred    = self.model(batch_x.permute(0, 2, 1)).permute(0, 2, 1) # (B, L_y, pred_len)
         y_pred = y_pred + trends_y  # (B, L_y, pred_len)
         return y_pred
 if __name__ == '__main__':
